export exif orientation
this builds on https://github.com/opencv/opencv/pull/8492 and
https://github.com/opencv/opencv/pull/8511 to add a new method,
.orientation(), to ImageDecoder. The JPEG subclass now overrides this to
do jpeg-specific exif tag reading

we also now export a method, OrientationTransform, which uses this
orientation value to do the actual matrix operations
