+++ modules/imgcodecs/src/loadsave.cpp
@@ -46,59 +46,13 @@
 #include "precomp.hpp"
 #include "grfmts.hpp"
 #include "utils.hpp"
-#include "exif.hpp"
 #undef min
 #undef max
 #include <iostream>
-#include <fstream>
 
 /****************************************************************************************\
 *                                      Image Codecs                                      *
 \****************************************************************************************/
-namespace {
-
-class ByteStreamBuffer : public std::streambuf
-{
-public:
-   ByteStreamBuffer(char* base, size_t length)
-   {
-      setg( base, base, base + length );
-   };
-
-protected:
-   virtual pos_type seekoff( off_type off,
-                             std::ios_base::seekdir dir,
-                             std::ios_base::openmode )
-   {
-      if( off > INT_MAX )
-      {
-         return -1;
-      }
-
-      if( off < INT_MIN )
-      {
-         return -1;
-      }
-
-      if( dir == std::ios_base::cur )
-      {
-         if( gptr() + off >= egptr() )
-         {
-            return -1;
-         }
-         if( gptr() + off < eback() )
-         {
-            return -1;
-         }
-         gbump( static_cast<int>(off) );
-      }
-
-      return gptr() - eback();
-   }
-};
-
-}
-
 namespace cv
 {
 
@@ -277,7 +231,7 @@ Ptr<ImageEncoder> findEncoder( const String& _ext )
 
 enum { LOAD_CVMAT=0, LOAD_IMAGE=1, LOAD_MAT=2 };
 
-static void ExifTransform(int orientation, Mat& img)
+void OrientationTransform(ImageOrientation orientation, Mat& img)
 {
     switch( orientation )
     {
@@ -312,50 +266,6 @@ static void ExifTransform(int orientation, Mat& img)
             //by default the image read has normal (JPEG_ORIENTATION_TL) orientation
             break;
     }
-}
-
-static void ApplyExifOrientation(const String& filename, Mat& img)
-{
-    int orientation = IMAGE_ORIENTATION_TL;
-
-    if (filename.size() > 0)
-    {
-       std::ifstream stream( filename.c_str(), std::ios_base::in | std::ios_base::binary );
-       ExifReader reader( stream );
-       if( reader.parse() )
-       {
-           ExifEntry_t entry = reader.getTag( ORIENTATION );
-           if (entry.tag != INVALID_TAG)
-           {
-               orientation = entry.field_u16; //orientation is unsigned short, so check field_u16
-           }
-       }
-       stream.close();
-    }
-
-    ExifTransform(orientation, img);
-}
-
-static void ApplyExifOrientation(const Mat& buf, Mat& img)
-{
-    int orientation = IMAGE_ORIENTATION_TL;
-
-    if( buf.isContinuous() )
-    {
-       ByteStreamBuffer bsb( reinterpret_cast<char*>(buf.data), buf.total() * buf.elemSize() );
-       std::istream stream( &bsb );
-       ExifReader reader( stream );
-       if( reader.parse() )
-       {
-           ExifEntry_t entry = reader.getTag( ORIENTATION );
-           if (entry.tag != INVALID_TAG)
-           {
-               orientation = entry.field_u16; //orientation is unsigned short, so check field_u16
-           }
-       }
-    }
-
-    ExifTransform(orientation, img);
 }
 
 /**
@@ -465,6 +375,12 @@ imread_( const String& filename, int flags, int hdrtype, Mat* mat=0 )
         return 0;
     }
 
+    /// optionally rotate the data if EXIF' orientation flag says so
+    if( (flags & IMREAD_IGNORE_ORIENTATION) == 0 && flags != IMREAD_UNCHANGED )
+    {
+        OrientationTransform(decoder->orientation(), *mat);
+    }
+
     if( decoder->setScale( scale_denom ) > 1 ) // if decoder is JpegDecoder then decoder->setScale always returns 1
     {
         resize( *mat, *mat, Size( size.width / scale_denom, size.height / scale_denom ) );
@@ -535,7 +451,7 @@ imreadmulti_(const String& filename, int flags, std::vector<Mat>& mats)
             // optionally rotate the data if EXIF' orientation flag says so
             if( (flags & IMREAD_IGNORE_ORIENTATION) == 0 && flags != IMREAD_UNCHANGED )
             {
-                ApplyExifOrientation(filename, mat);
+                OrientationTransform(decoder->orientation(), mat);
             }
 
             break;
@@ -566,12 +482,6 @@ Mat imread( const String& filename, int flags )
 
     /// load the data
     imread_( filename, flags, LOAD_MAT, &img );
-
-    /// optionally rotate the data if EXIF' orientation flag says so
-    if( (flags & IMREAD_IGNORE_ORIENTATION) == 0 && flags != IMREAD_UNCHANGED )
-    {
-        ApplyExifOrientation(filename, img);
-    }
 
     /// return a reference to the data
     return img;
@@ -671,6 +581,7 @@ imdecode_( const Mat& buf, int flags, int hdrtype, Mat* mat=0 )
     CvSize size;
     size.width = decoder->width();
     size.height = decoder->height();
+    ImageOrientation orientation = decoder->orientation();
 
     int type = decoder->type();
     if( (flags & IMREAD_LOAD_GDAL) != IMREAD_LOAD_GDAL && flags != IMREAD_UNCHANGED )
@@ -723,6 +634,12 @@ imdecode_( const Mat& buf, int flags, int hdrtype, Mat* mat=0 )
         return 0;
     }
 
+    /// optionally rotate the data if EXIF' orientation flag says so
+    if( (flags & IMREAD_IGNORE_ORIENTATION) == 0 && flags != IMREAD_UNCHANGED )
+    {
+        OrientationTransform(orientation, *mat);
+    }
+
     return hdrtype == LOAD_CVMAT ? (void*)matrix :
         hdrtype == LOAD_IMAGE ? (void*)image : (void*)mat;
 }
@@ -733,12 +650,6 @@ Mat imdecode( InputArray _buf, int flags )
     Mat buf = _buf.getMat(), img;
     imdecode_( buf, flags, LOAD_MAT, &img );
 
-    /// optionally rotate the data if EXIF' orientation flag says so
-    if( (flags & IMREAD_IGNORE_ORIENTATION) == 0 && flags != IMREAD_UNCHANGED )
-    {
-        ApplyExifOrientation(buf, img);
-    }
-
     return img;
 }
 
@@ -747,12 +658,6 @@ Mat imdecode( InputArray _buf, int flags, Mat* dst )
     Mat buf = _buf.getMat(), img;
     dst = dst ? dst : &img;
     imdecode_( buf, flags, LOAD_MAT, dst );
-
-    /// optionally rotate the data if EXIF' orientation flag says so
-    if( (flags & IMREAD_IGNORE_ORIENTATION) == 0 && flags != IMREAD_UNCHANGED )
-    {
-        ApplyExifOrientation(buf, *dst);
-    }
 
     return *dst;
 }
